requirements.txt
faiss-cpu
flask
flask-cors
sentence-transformers
numpy
pandas
scikit-learn
